Implemented start countdown
Start countdown works by disabling the car controller script on the car and hiding the racing UI on entry to the scene. Once the timer is over both the script and race UI are enabled starting the race timer as well. Only catch to this approach the countdown canvas must be separate from the Race UI object in the scene.